[feature] Remove Any CPU from pinvoke calls (#57)
* [ignore] Remove Any CPU

* Use 127.0.0.1:0 as TCP local address

* Increased wait time for the GetDiscoveredTopicData call

* Use TCP transport for the TestGetDiscoveredTopicData

* Use other participant to discover topics data

* Use localhost as TCP local address

* Remove LocalAddress from TCP transport

* Assign port to the local/public address

* Use StatusMask::NoStatusMask when removing listener

* Remove debug parameters

* Use TCP transport on the discovery topic tests

* Explicitly clean up entities

* Use default QoS in TestGetDiscoveredTopicData

* Create reader/writer in TestGetDiscoveredTopicData test

* Fix BuiltinTopicKey type (byte[16])

* Fixed platform environment variables

* Adde try-finally blocks

* Added Reader/Writer to the TestGetDiscoveredTopics

* Added PublicAddress to TCP transport

* Removed PublicAddress

* Changed to localhost

* Added parameters to the DCPSInfoRepo process

* Try without transport binding (default)

* Removed WaitForSubscription/Publication checks

* Using UDP transport

* Ignore TopicData test until fix

+++ Sources/OpenDDSharp/BuiltinTopicKey.cpp
@@ -19,12 +19,12 @@ along with OpenDDSharp. If not, see <http://www.gnu.org/licenses/>.
 **********************************************************************/
 #include "BuiltinTopicKey.h"
 
-array<System::Int32, 1>^ OpenDDSharp::DDS::BuiltinTopicKey::Value::get() {
+array<System::Byte, 1>^ OpenDDSharp::DDS::BuiltinTopicKey::Value::get() {
 	return m_value;
 };
 
 void OpenDDSharp::DDS::BuiltinTopicKey::FromNative(::DDS::BuiltinTopicKey_t native) {
-	m_value = gcnew array<System::Int32, 1>(16);
+	m_value = gcnew array<System::Byte, 1>(16);
 	for (int i = 0; i < 16; i++) {
 		m_value[i] = native.value[i];
 	}
